Add C99 compatibility for for-loops
The `decl` rule is split to a `decl_body` sub-rule which does not
contain `';'` at the end, to enable declarations in the for-loop
init body.

Following example could not be parsed with the old rule:

```
for (int i = 0;;)
```

The parsing failed because `int i = 0;` is reduced to `decl`, but
the semi-column at the end shoud not be part of the `decl`, but a
regular part of the for-loop.

Because of this, `decl_body` is used inside a `iteration_statement`
rule, instead of `decl`.