Add disambiguation solvers and refactor code
There were two productions in grammar that, in most cases, caused
disambiguation. Those are:

```
typedef_name:      ID;
direct_declarator: ID;
```

The problem is solved as follows. Sematic action is called for
every `decl_body` production where the names user-defined types
(struct, union, typedef) are collected. Dynamic filter is registered
for the `typedef_name` production where we check if the `ID` that
is trying to be reduced is in the collection of user-defined types.
If so, then the reduction is performed. Otherwise, we reject the
reduction.